fix(clippy): auto-fix format strings and deprecated IndexMap::remove
Address clippy warnings by:
- Prefixing unused variables with underscore
- Adding #[allow(dead_code)] for intentionally unused helper functions
- Using div_ceil() instead of manual ceiling division
- Simplifying map_or() to is_some_and()
- Fixing type complexity issues with type aliases
- Using .copied() instead of .map(|k| *k)
- Fixing digit grouping inconsistencies (3_600_000)
- Adding #[allow(non_snake_case)] for Meilisearch API-compatible structs
- Removing unnecessary casts
- Fixing await_holding_lock issues

Closes: bf-66nh

+++ crates/miroir-core/src/dump_chunking.rs
@@ -262,7 +262,7 @@ mod tests {
         let lines_per_chunk = (256 * 1024 * 1024) / line_size;
         let total_lines = lines_per_chunk * 4; // 4 chunks
 
-        let data = create_test_data(total_lines as usize, line_size - 20);
+        let data = create_test_data(total_lines, line_size - 20);
         let chunks = split_dump_into_chunks(&data, 256 * 1024 * 1024);
 
         // Should get approximately 4 chunks
